source_code/backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
bcrypt==3.2.2
boto3==1.40.29
botocore==1.40.29
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.6
django-cors-headers==4.3.0
django-environ==0.8.1
django-filter==23.1
django-redis==6.0.0
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.14.0
drf-spectacular==0.28.0
drf-yasg==1.21.0
gevent==23.9.1
greenlet==3.2.4
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.4
MarkupSafe==2.1.2
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==23.0
pdfkit==1.0.0
pycparser==2.21
pymongo==3.13.0
pyparsing==3.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
hiredis==2.4.0
py-redis==1.1.1
redis-om==0.2.1
redis-search-django==0.1.0
types-redis==4.6.0.20241004
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
setuptools==80.9.0
simplejson==3.18.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.5.0
websocket==0.2.1
wheel==0.45.1
zipp==3.12.0
zope.event==5.0
zope.interface==6.0
pydantic==1.10.24